[iceberg] Add UUID type support
The iceberg spec lists uuid as a valid schema type. Presto supports
UUID types but there was no support for reading or writing them
in the connector.

This commit makes the necessary changes in the connector to create
tables with UUID columns and support for UUIDs in the parquet reader.
This includes an implementation for UUIDs in the batchreader.